% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairingGame.R
\name{pairingGame}
\alias{pairingGame}
\title{Play Pairing Game}
\usage{
pairingGame(a, b, strata = pickerFirst, stratb = pickerFirst)
}
\arguments{
\item{a}{numeric vector of player a pieces}

\item{b}{numeric vector of player b pieces of length a}

\item{strata}{picker function to select between two numbers for player a}

\item{stratb}{picker function to select between two numbers for player b}
}
\value{
length 2 vector number of wins for player a and player b
}
\description{
Play a game by providing a sequence of numbers (typically 5).
The game is played in the following sequence:\enumerate{
    \item player a reveals her first number
    \item player b reveals his first two numbers
    \item player a selects one of player b's revealed numbers to match with her revealed number
    \item player a reveals her next two numbers
    \item player b selects one of player a's revealed numbers to match with his revealed number
    \item player b reveals his next two numbers
    \item player a selects one of player b's revealed numbers to match with her revealed number
    \item player a reveals her next two numbers
    \item player b selects one of player a's revealed numbers to match with his revealed number
    \item the remaining two numbers are matched.
}

For each match, the player with the higher number wins one point.
The player with the most points wins the game.
The picker functions must have arguments 
active, revealed, type and exclude.
active and revealed must be numeric vectors. 
type must be a length 1 character vector with value "reveal" or "pair".
If type is "reveal", the function must pick the index of active to reveal.
If type is "pair", the function must pick index of revealed to pair.
exclude must be NA, or a numeric vector, 
which is the index of revealed to ignore when type is "reveal".
}
\examples{
pairingGame(a = 1:5, b = 5:1)
pairingGame(a = 10:14, b = 10:14)
pairingGame(a = 1:5, b = c(5, 1:4))
pairingGame(a = rep(1, 5), b = rep(1, 5), 
    strata = pickerFirst, stratb = pickerFirst)
pairingGame(a = 1:5, b = 5:1, strata = pickerRandom)
pairingGame(a = 1:5, b = c(5, 1:4), 
    strata = pickerRandom, stratb = pickerRandom)
pairingGame(a = 1:5, b = c(5, 1:4), 
    strata = pickerRandom, stratb = pickerMax)
pairingGame(a = 1:5, b = c(5, 1:4), 
    strata = pickerRandom, stratb = pickerJustMax)
set.seed(22532)
pairingGame(a = sample(1:5, size = 5), b = sample(1:5, size = 5), 
    strata = pickerRandom, stratb = pickerRandom)
}

