% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickers.R
\name{pickerFirst}
\alias{pickerFirst}
\title{Pick First Match Presented}
\usage{
pickerFirst(active, revealed, exclude = NA, type = c("pair", "reveal"))
}
\arguments{
\item{active}{numeric vector choices of active player}

\item{revealed}{numeric vector choices revealed by non-active player for next round}

\item{exclude}{integer vector of indices of active 
to exclude when picking active for type reveal.}

\item{type}{single character, select modality:\itemize{
    \item "pair", pick index of revealed to create pairing.
    \item "reveal", pick index of active to reveal.
}}
}
\value{
single numeric
}
\description{
Select index of first active or revealed.
If type is "reveal", pick index of active to reveal (always first).
If type is "pair", pick index of revealed to pair (always first).
}
\examples{
pickerFirst(revealed = 1:2)
pickerFirst(revealed = 1)
pickerFirst(active = rep(1, 5), 
    revealed = rep(1, 5), type = "reveal")
pickerFirst(active = 1:5, 
    revealed = 1, exclude = 1, type = "reveal")
}

