% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickers.R
\name{pickerRandom}
\alias{pickerRandom}
\title{Pick a Match at Random}
\usage{
pickerRandom(active, revealed, exclude = NA, type = c("pair", "reveal"))
}
\arguments{
\item{active}{numeric vector choices of active player}

\item{revealed}{numeric vector choices revealed by non-active player for next round}

\item{exclude}{integer vector of indices of active 
to exclude when picking active for type reveal.}

\item{type}{single character, select modality:\itemize{
    \item "pair", pick index of revealed to create pairing.
    \item "reveal", pick index of active to reveal.
}}

\item{prob}{single numeric between 0 and 1}
}
\value{
single numeric
}
\description{
Selects index of random active or revealed.
If type is "reveal", pick index of 
active to reveal (always random).
If type is "pair", pick index of revealed to pair (always random).
}
\examples{
pickerRandom(revealed = 3:4)
pickerRandom(revealed = 5)
pickerRandom(active = 1:5, revealed = 1:5, type = "reveal")
pickerRandom(active = 2, revealed = 2, exclude = 1)
}

